requirements.txt
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
databases==0.7.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.89.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pydantic==1.10.4
python-decouple==3.7
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
zipp==3.14.0